requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.1
Django==5.1.4
django-allauth==65.4.1
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psutil==6.1.1
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
